requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langgraph==0.2.28
sentence-transformers==3.0.1
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
tqdm==4.66.5